+++ requirements.txt
@@ -11,3 +11,4 @@ asyncio>=3.4,<4.0
 aiohttp>=2.0,<4.0
 async_timeout>=2.0,<4.0
 fuzzywuzzy[speedup]
+git+https://github.com/Rapptz/discord-ext-menus
